implemented deep q agent. Dealing with compile issues, need to separate files

+++ simulations/CartPoleSimulation.h
@@ -16,7 +16,7 @@ public:
     void eval(int tid, int block, float ** actions, float * gamestate);
     int checkFinished(int tid, int block, float * gamestate);
     void setOutput(int tid, int block, float * output, float * gamestate, const float * startingParams_d);
-    Eigen::MatrixXd getState(float & action, float & reward, float *gamestate);
+    Eigen::MatrixXd getState(int & action, float & reward, float *gamestate);
     int getID();
     
 };
