+++ simulations/TargetSimulation.h
@@ -24,7 +24,7 @@ public:
 
     void setOutput(int tid, int block, float * output, float * gamestate, const float * startingParams_d);
 
-    Eigen::MatrixXd getState(float & action, float & reward, float *gamestate);    
+    Eigen::MatrixXd getState(int & action, float & reward, float *gamestate);    
     
     int getID();
